Basic Gradle layout compiles
