+++ build.gradle
@@ -1,0 +1,12 @@
+apply plugin: 'java'
+
+sourceCompatibility = 1.7
+version = '1.0.5'
+
+jar {
+    manifest {
+        attributes(
+                'Main-Class': 'cdb.dump'
+        )
+    }
+}
