+++ src/main/java/com/strangegizmo/cdb/CdbMake.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2000-2001, Michael Alyn Miller <malyn@strangeGizmo.com>
+ * Copyright (c) 2000-2001, Michael Alyn Miller &lt;malyn@strangeGizmo.com&gt;
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -40,7 +40,7 @@ import java.util.*;
  * CdbMake implements the database-creation side of
  * D.&nbsp;J.&nbsp;Bernstein's constant database package.
  *
- * @author		Michael Alyn Miller <malyn@strangeGizmo.com>
+ * @author		Michael Alyn Miller &lt;malyn@strangeGizmo.com&gt;
  * @version		1.0.0
  */
 public final class CdbMake {
